Remove all seqs stuff and leave only scrabble
